requirements.txt
pytest>=8.0.0
httpx>=0.27.0
pydantic>=2.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0

# 测试报告和并行执行
pytest-html>=3.2.0
pytest-xdist>=3.0.0
pytest-rerunfailures>=10.0
pytest-cov>=4.0.0
allure-pytest>=2.9.0

# UI测试
playwright>=1.45.0
pytest-playwright>=0.5.0

# API测试增强
requests>=2.25.0

# 数据驱动测试
pytest-datadir>=1.3.1

